Make the installer build script portable
The installer script was previously not portable, so you couldn't build it
unless your directory structure was exactly the same as Lixkote's, even
down to having the same Windows user name. This makes the installer build
script use relative paths so it works no matter where the source files are
located.

Signed-off-by: Isabella <kawapure@gmail.com>