Refactor TileColorCalculation.cs
I refactored the code in TileColorCalculation.cs to be a lot more readable.
The per-application overrides were separated from the colour string checks,
and the colour string checks were moved to a switch to be more readable.

I wish this project used a newer version of C#, because switch expressions
are a lot nicer, but this is already much more readable and maintainable
than the previous version.

Per-application overrides, which are done by the program name, have been
moved to use anonymous string arrays via LINQ. This new pattern involves
less repetition and thus results in shorter lines, so I decided to go with
it.

There is one breaking change with this: previously, the colour names were
only compared to certain case variants (i.e. "lightgreen", "Lightgreen",
and "LightGreen"). In favour of code readability, any case is now allowed
as the main string is converted to lowercase and then compared with a
lowercase string constant. I don't personally view this to be any cause of
trouble, but it's worth pointing out where the behaviour differs
regardless.

Signed-off-by: Isabella <kawapure@gmail.com>