Port DirectHook to C#
...sort of. I really rewrote the whole thing, if that's okay. There was a
lot of functionality from the AHK implementation I wrote off as boilerplate
and decided to remove.

This basically just adds C interop interfaces for the native code the AHK
script calls into. No abstractions are implemented yet. Most notably, the
management of events is currently delegated to the users of them, rather
than added to a static collection and managed by the DirectHook module.
This is a rather C-like approach which should be cleaned up in the future,
but the groundwork implemented here is good regardless.

The rest of the project is already implemented in C#, so I didn't see any
reason for there to be an external process that manages events over a named
pipe instead of just implementing it in C#. Especially if it's being done
in a much slower runtime environment like that of AutoHotKey.

For now, all interfaces are marked as internal. These should probably be
changed to private if/when an abstraction is implemented, but for now, the
C foreign function interface is the only way to interact with this API.

Signed-off-by: Isabella <kawapure@gmail.com>